requirements.txt
pudb
linux-metrics
kafka-python
psycopg2-binary
